refactored file directories and deleted unused files
