+++ components/Layout/index.js
@@ -1,4 +1,4 @@
-// @author Charles Liu
+import PropTypes from 'prop-types';
 
 import Meta from './Meta';
 
@@ -6,9 +6,13 @@ const Layout = ({ children }) => {
   return (
     <>
       <Meta />
-      <main className="">{children}</main>
+      <main>{children}</main>
     </>
   );
+};
+
+Layout.propTypes = {
+  children: PropTypes.node.isRequired,
 };
 
 export default Layout;
