+++ pages/dashboard/index.js
@@ -2,10 +2,11 @@ import React from 'react';
 import { useState } from 'react';
 import PropTypes from 'prop-types';
 import { getSession } from 'next-auth/client';
+import classNames from 'classnames';
+import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
+
 import { Grid, Hidden, Drawer, Fab } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
-import classNames from 'classnames';
-
 import MenuIcon from '@material-ui/icons/Menu';
 import IconButton from '@material-ui/core/IconButton';
 import SettingsIcon from '@material-ui/icons/Settings';
@@ -145,7 +146,7 @@ Dashboard.propTypes = {
   friendData: PropTypes.array.isRequired,
 };
 
-export const getServerSideProps = async ({ req }) => {
+export const getServerSideProps = async ({ req, locale }) => {
   const session = await getSession({ req });
 
   await dbConnect();
@@ -162,6 +163,7 @@ export const getServerSideProps = async ({ req }) => {
   return {
     props: {
       session: JSON.stringify(newSession) ?? '', // otherwise nextjs throws error - can't serialize data
+      ...(await serverSideTranslations(locale, ['common', 'dashboard'])),
       friendData: [
         {
           name: 'Yi Nan Zhang',
