+++ pages/friends/index.js
@@ -9,7 +9,7 @@ function Users() {
   const [session, setSession] = useState();
   const [currentUser, setCurrentUser] = useState({});
 
-  useEffect(async () => {
+  const init = async () => {
     const userSession = await getSession();
     setSession(userSession);
     if (userSession) {
@@ -19,7 +19,7 @@ function Users() {
         email: userSession.user.email,
       });
 
-      const res = await axios.get('/api/getuserlist', {
+      const res = await axios.get('/api/user/get-user-list', {
         params: {
           name: userSession.user.name,
           email: userSession.user.email,
@@ -28,11 +28,15 @@ function Users() {
       setUserList(res.data.users);
       setFriendsList(res.data.friends);
     }
+  };
+
+  useEffect(() => {
+    init();
   }, []);
 
   async function handleAddFriend(user) {
     console.debug('[handleAddFriend] send');
-    const result = await axios.post('/api/createnewfriend', {
+    const result = await axios.post('/api/user/create-new-friend', {
       user1name: currentUser.name,
       user1email: currentUser.email,
       user2name: user.name,
@@ -46,15 +50,11 @@ function Users() {
     <>
       <h1> Users </h1>
       {userList.map((user, i) => {
-        if (user.name == currentUser.name && user.email == currentUser.email)
-          return;
+        if (user.name == currentUser.name && user.email == currentUser.email) return;
         let friendText = 'Add';
         for (let friend in friendsList) {
           friend = friendsList[friend];
-          if (
-            friend.requester_email == currentUser.email &&
-            friend.recipient_email == user.email
-          ) {
+          if (friend.requester_email == currentUser.email && friend.recipient_email == user.email) {
             if (friend.status == 1) {
               friendText = 'Pending';
               break;
@@ -63,10 +63,7 @@ function Users() {
               break;
             }
           }
-          if (
-            friend.recipient_email == currentUser.email &&
-            friend.requester_email == user.email
-          ) {
+          if (friend.recipient_email == currentUser.email && friend.requester_email == user.email) {
             if (friend.status == 1) {
               friendText = 'Accept';
               break;
@@ -81,10 +78,7 @@ function Users() {
           <div key={i}>
             {user.name} {user.email}
             <img src={user.image} alt="user" />
-            <Button onClick={() => handleAddFriend(user)}>
-              {' '}
-              {friendText}{' '}
-            </Button>
+            <Button onClick={() => handleAddFriend(user)}> {friendText} </Button>
           </div>
         );
       })}
